Create README.md
# WorldCryptoNetwork Manifesto
==================

## A decentralised semi-autonomous open source independent news organisation

### Article 1
Tell the truth, tell it early, tell it to the people who need to hear it the most.

### Article 2
Be real, we don’t do fake around here. That means just tell shit how it really is! 

### Article 3
Respect your sources, credit that Redditor who broke the story, hell link her thread in the description and give her some coins!

### Article 4
Respect your audience. They are giving you their time that they are never getting back. Make them appreciate its value by helping them gain insights, new perspectives and never assume like they can’t teach you something in return.

Journalism is a dialogue. It’s not about talking down at your audience as if your version of the truth is the only one. *Your audience are on the front line!*

The more you respect the time of your audience the more they will respect it themselves.

The more intelligent you treat them the more intelligent they will feel and act.

### Article 5
Remember [Cunninghams’s law](https://meta.wikimedia.org/wiki/Cunningham's_Law):
"the best way to get the right answer on the Internet is not to ask a question, it's to post the wrong answer."

Which means put in a pull request and change this document the way you believe it needs to be.

### Article 6
Take action
Comment, retweet, reddit and make your own videos or podcasts